pyproject.toml
[tool.poetry]
name = "stegstash"
version = "2020"
description = "Provides multiple methods to hide and retrieve data"
authors = ["FredHappyface"]
classifiers = [
	"Environment :: Console",
	"Environment :: MacOS X",
	"Environment :: Win32 (MS Windows)",
	"Development Status :: 5 - Production/Stable",
	"Intended Audience :: Developers",
	"Intended Audience :: Education",
	"License :: OSI Approved :: MIT License",
	"Natural Language :: English",
	"Operating System :: OS Independent",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3.5",
	"Programming Language :: Python :: 3.6",
	"Programming Language :: Python :: 3.7",
	"Programming Language :: Python :: 3.8",
	"Programming Language :: Python :: Implementation :: CPython",
	"Topic :: Security",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Utilities"
]
homepage = "https://github.com/FHPythonUtils/StegStash"
repository = "https://github.com/FHPythonUtils/StegStash"
documentation = "https://github.com/FHPythonUtils/StegStash/blob/master/README.md"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.5"
soundfile = "^0.10.3"
numpy = "^1.18.4"
metprint = "^2020.6.1"
pillow = "^7.1.2"

[tool.dephell.main]
from = "pyproject.toml"
to = "requirements.txt"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
